Removed old UserActivity.
