+++ app/src/main/AndroidManifest.xml
@@ -22,14 +22,15 @@
             android:name="firebase_analytics_collection_enabled"
             android:value="false"/>
 
-        <activity
-            android:name=".user.UserActivity"
-            android:theme="@style/AppTheme.Transparent.Dark"
-            android:label="@string/app_name">
+        <activity android:name=".user.UserActivity"
+                  android:theme="@style/AppTheme.Transparent.Dark"
+                  android:label="@string/app_name">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
+
         </activity>
 
         <activity
@@ -79,16 +80,6 @@
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
             </intent-filter>
-        </activity>
-
-        <activity android:name=".user.UserActivity2"
-                  android:theme="@style/AppTheme.Transparent.Dark"
-                  android:exported="true">
-
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-            </intent-filter>
-
         </activity>
 
         <activity android:name=".editors.IssueEditor"
